requirements.txt
psutil
tkinterweb
pillow
tkintermapview
